@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import { BrowserRouter as Router } from "react-router-dom";
 import "./index.css";
 import App from "./App";
@@ -8,15 +8,17 @@ import AppProvider from "./context";
 import "bootstrap/dist/css/bootstrap.min.css";
 import { auth } from "./firebase";
 
-ReactDOM.render(
+const rootElement = document.getElementById("root");
+const root = createRoot(rootElement); // Create root using createRoot
+
+root.render(
 	<React.StrictMode>
 		<Router>
 			<AppProvider auth={auth}>
 				<App />
 			</AppProvider>
 		</Router>
-	</React.StrictMode>,
-	document.getElementById("root")
+	</React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
