@@ -1,2 +1,2 @@
 // export const URL_FOR_API = "http://localhost:8000";
-export const URL_FOR_API = "https://estudia-ebau-api.herokuapp.com";
+export const URL_FOR_API = "https://my-vm-jurko.eastus2.cloudapp.azure.com/school-study";
